`read_command`: check if command too long
Modify `read_command` function in `shell.c` for safety, it would now
check if the command entered is too long, if so, the function would
soonly return and the command wouldn't be parsed.
Also, the `reset` and `cancel_reset` functions have been seperated from
the `shell.c`, and now in `reboot.c` for the easier management.