[Lab4] Get usable memory region from device tree
callback function `find_root_node` tries to find the root node in the
dtb, and get the `#address-cells` and `size-cells` property, which
specifies the number of <u32> cells to represent the address and size in the
`reg` property in the children of root. With these two value, we can use
the callback function `find_memory_node`, which tries to find the usable
memroy region specified in the `reg` property in the memory node, this
function get the usable memory's start address and length, for now, this
function can only work when there's only one memory node, if not, the
address and length would be covered by the last memory node.