Correct reserve page address
In function `register_reserve_pages`, the end address should not be
included as the reserved one, instead, use the `end - PAGE_SIZE` as the
start of the final page.
In function `buddy_init`, the `register_reserve_pages` function call to
reserve from usable memory end to the end of the power-of-two buddy
system maximum address, the end address is wrong, since the `MAX_ALLOC2`
is the maximum number of pages, so it should be multiply by PAGE_SIZE to
get the actual size.