[Lab4] Fix the bug not properly free object
When freeing an object from a slab which size is 1 page (Not a compound
page by default), so this line `if (!PageHead(page)) return;` will be
the problem, thus we remove this line. Another problem is that when we
call `kmalloc` with a large size that is not affordable by slab
allocator, it would directly request it from page allocator, so when we
want to free the object we allocated previously, we would call
`page_free` function, which should receive a page address and a order as
parameters. The problem is in `kfree`, the `compound_order` we get is
wrong, since we do not store the compound order in the PageHead, instead,
we store it in the first PageTail.